refactor tools to shared util
